silence a couple of Flexelint complaints.
git-svn-id: http://www.varnish-cache.org/svn/trunk/varnish-cache@3548 d4fa192b-c00b-0410-8231-f00ffab90ce4

+++ lib/libvarnish/tcp.c
@@ -216,8 +216,8 @@ void
 TCP_set_read_timeout(int s, double seconds)
 {
 	struct timeval timeout;
-	timeout.tv_sec = floor(seconds);
-	timeout.tv_usec = 1e6 * (seconds - timeout.tv_sec);
+	timeout.tv_sec = (int)floor(seconds);
+	timeout.tv_usec = (int)(1e6 * (seconds - timeout.tv_sec));
 #ifdef SO_RCVTIMEO_WORKS
 	AZ(setsockopt(s, SOL_SOCKET, SO_RCVTIMEO, &timeout, sizeof timeout));
 #endif
